(#12498) Add ext to puppet-common package
The ext folder has useful files in it for puppet users and is already installed
in puppet rpms. This patch adds the ext to the puppet-common package, to be
installed in the /usr/share/puppet directory. It also add this directory to the
puppet-common.dirs list and adds a tag to be overridden to
puppet-common.lintian-overrides.