Enable apache2 headers module on passenger install
The apache conf that ships with passenger was updated recently to include some
RequestHeaders directives, which require the headers module to be enabled for
apache. This patch enables the headers module in the postinst, before apache is
restarted. Without this patch, if the headers module is not installed, apache
won't start, so the postinst for the puppetmaster package fails.