(#14080) Add suggests of libactiverecord-ruby and ruby-activerecord
Puppetmaster already has a suggests of rails, which is how one comes by
activerecord in ubuntu circa lucid. That changed for ubuntu circa precise, when
ruby-activerecord and libactiverecord-ruby were available as standalone
packages, but also brought in as dependencies of rails. This commit modifies
the suggestion of rails to include ruby-activerecord and libactiverecord-ruby
as viable options to all of rails.