Remove puppetmaster.logrotate due to redundancy
This commit removes the logrotate configuration for
puppetmaster and effectively combines it with puppet's.
This is necessary because duplicate log entry errors were
thrown when both puppet and puppetmaster were in use. Files
to be rotated in puppetmaster's configuration are already
covered by puppet's configuration, and thus it is not needed.

Conflicts:

	debian/changelog